Move navigation modifier code to navigation.js file
